refactor(sdk): release lock after discovery complete
Add a wrapper function to call discovery.Discovery() and release the
lock when call completed so that developer will not be forced to push
something to deviceCh to release the lock

fix #609
